+++ Dissimilis.WebAPI/Database/Interfaces/IBar.cs
@@ -2,7 +2,7 @@
 {
     public interface IBar : IBaseEntity
     {
-        byte BarNumber { get; }
+        ushort BarNumber { get; }
         byte? House { get; }
         int Id { get; }
         Part Part { get; }
